flask_api/requirements.txt
alembic==1.10.3
bcrypt==4.0.1  
click==8.1.3   
colorama==0.4.6
DateTime==5.1  
distlib==0.3.6
filelock==3.12.0
Flask==2.2.3
Flask-Alembic==2.0.1
Flask-DotEnv==0.1.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
platformdirs==3.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.9
typing_extensions==4.5.0
virtualenv==20.21.0
Werkzeug==2.2.3
zope.interface==6.0